Non visual for RasterizerState
