+++ Test/Framework/Visual/RasterizerStateTest.cs
@@ -13,62 +13,6 @@ namespace MonoGame.Tests.Visual
     [TestFixture]
     internal class RasterizerStateTest : VisualTestFixtureBase
     {
-        [Test]
-        public void ShouldNotBeAbleToSetNullRasterizerState()
-        {
-            Game.DrawWith += (sender, e) =>
-            {
-                Assert.Throws<ArgumentNullException>(() => Game.GraphicsDevice.RasterizerState = null);
-            };
-            Game.Run();
-        }
-
-        [Test]
-        public void ShouldNotBeAbleToMutateStateObjectAfterBindingToGraphicsDevice()
-        {
-            Game.DrawWith += (sender, e) =>
-            {
-                var rasterizerState = new RasterizerState();
-
-                // Can mutate before binding.
-                DoAsserts(rasterizerState, Assert.DoesNotThrow);
-
-                // Can't mutate after binding.
-                Game.GraphicsDevice.RasterizerState = rasterizerState;
-                DoAsserts(rasterizerState, d => Assert.Throws<InvalidOperationException>(d));
-
-                // Even after changing to different RasterizerState, you still can't mutate a previously-bound object.
-                Game.GraphicsDevice.RasterizerState = RasterizerState.CullCounterClockwise;
-                DoAsserts(rasterizerState, d => Assert.Throws<InvalidOperationException>(d));
-            };
-            Game.Run();
-        }
-
-        [Test]
-        public void ShouldNotBeAbleToMutateDefaultStateObjects()
-        {
-            Game.DrawWith += (sender, e) =>
-            {
-                DoAsserts(RasterizerState.CullClockwise, d => Assert.Throws<InvalidOperationException>(d));
-                DoAsserts(RasterizerState.CullCounterClockwise, d => Assert.Throws<InvalidOperationException>(d));
-                DoAsserts(RasterizerState.CullNone, d => Assert.Throws<InvalidOperationException>(d));
-            };
-            Game.Run();
-        }
-
-        private static void DoAsserts(RasterizerState rasterizerState, Action<TestDelegate> assertMethod)
-        {
-            assertMethod(() => rasterizerState.CullMode = CullMode.CullClockwiseFace);
-            assertMethod(() => rasterizerState.DepthBias = 0.1f);
-            assertMethod(() => rasterizerState.FillMode = FillMode.WireFrame);
-            assertMethod(() => rasterizerState.MultiSampleAntiAlias = true);
-            assertMethod(() => rasterizerState.ScissorTestEnable = true);
-            assertMethod(() => rasterizerState.SlopeScaleDepthBias = 0.2f);
-#if !XNA
-            assertMethod(() => rasterizerState.DepthClipEnable = false);
-#endif
-        }
-
         [TestCase(CullMode.CullClockwiseFace)]
         [TestCase(CullMode.CullCounterClockwiseFace)]
         [TestCase(CullMode.None)]
