+++ Test/MonoGame.Tests.XNA.csproj
@@ -125,6 +125,8 @@
     <Compile Include="Framework\Graphics\EffectTest.cs" />
     <Compile Include="Framework\Graphics\GraphicsAdapterTest.cs" />
     <Compile Include="Framework\Graphics\GraphicsDeviceTestFixtureBase.cs" />
+    <Compile Include="Framework\Graphics\OcclusionQueryNonVisualTest.cs" />
+    <Compile Include="Framework\Graphics\RasterizerStateNonVisualTest.cs" />
     <Compile Include="Framework\Graphics\SpriteFontNonVisualTest.cs" />
     <Compile Include="Framework\Graphics\Texture2DNonVisualTest.cs" />
     <Compile Include="Framework\Graphics\Texture3DNonVisualTest.cs" />
